캔들 데이터 저장 로직 구현
- 캔들 데이터 저장 로직 구현
- Candle Entiry 컬럼타입들  double로 변경
- 테스트 데이터 정리

+++ src/model/CandleEntity.ts
@@ -12,19 +12,19 @@ export class Candle {
   @PrimaryGeneratedColumn()
   id: number;
 
-  @Column("float")
+  @Column("double")
   open: number;
 
-  @Column("float")
+  @Column("double")
   high: number;
 
-  @Column("float")
+  @Column("double")
   low: number;
 
-  @Column("float")
+  @Column("double")
   close: number;
 
-  @Column("float")
+  @Column("double")
   volume: number;
 
   @CreateDateColumn({ name: "created_at" })
